fixing the product price type

+++ webshop/src/main/java/com/shop/webshop/models/Cart.java
@@ -36,7 +36,7 @@ public class Cart {
     public void updateTotal(){
         double total = 0;
         for(CartItem cartItem: this.cartItems){
-            total = total + (cartItem.getQuantity() * cartItem.getProduct().getPrice().doubleValue());
+            total = total + (cartItem.getQuantity() * cartItem.getProduct().getPrice());
         }
         this.cartTotal = total;
     }
@@ -71,7 +71,7 @@ public class Cart {
     public double getCartTotal() {
         cartTotal = 0;
         for(CartItem cartItem: this.cartItems){
-            cartTotal = cartTotal + (cartItem.getQuantity() * cartItem.getProduct().getPrice().doubleValue());
+            cartTotal = cartTotal + (cartItem.getQuantity() * cartItem.getProduct().getPrice());
         }
         return cartTotal;
     }
